Update index.blade.php
Laravel is buggy in the locale department, even though we specify a default locale of eng in app/

When you go into the container and run php tinker, both config('app.local'); and App::getLocale() are null -- if you go into the settings page, then set a language and save it, then it gets saved into the system default locale. This locale string is pulled automatically by the blade template and is null, so the value doesn't get populated which causes the login error about the language. Once the language is set, users can log in.